refactoring code for better practice and get build workig
