+++ src/app/signup/page.tsx
@@ -2,7 +2,7 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { createUserWithEmailAndPassword, onAuthStateChanged, GoogleAuthProvider, signInWithPopup } from "firebase/auth";
-import { auth } from "../firebaseconfig.js";
+import { auth } from "../../utils/firebaseconfig.js";
 import ResponsiveAppBar from "../navbar";
 import Link from 'next/link.js';
 import GoogleIcon from '@mui/icons-material/Google';
